fix: Fix after merge with 'cross-model-queries'
